Extract animation functions into  a separate file
