+++ src/app/animation.tsx
@@ -1,0 +1,77 @@
+export function animationDelay(i: number){
+  const num = i + 2 // avoid 1 and 0
+  const delay = generatePerlinNoise(num**2, num**2)[num**2] * 10 //i ** 2 % 4 / 10
+  // console.log(i, delay)
+  return delay
+}
+
+
+/* PERLIN NOISE */
+function generatePerlinNoise(width: number, height: number) {
+  let options = {};
+  var octaveCount =  4;
+  var amplitude =  0.1;
+  var persistence =  0.2;
+  var whiteNoise = generateWhiteNoise(width, height);
+
+  var smoothNoiseList = new Array(octaveCount);
+  var i;
+  for (i = 0; i < octaveCount; ++i) {
+    smoothNoiseList[i] = generateSmoothNoise(i);
+  }
+  var perlinNoise = new Array(width * height);
+  var totalAmplitude = 0;
+  // blend noise together
+  for (i = octaveCount - 1; i >= 0; --i) {
+    amplitude *= persistence;
+    totalAmplitude += amplitude;
+
+    for (var j = 0; j < perlinNoise.length; ++j) {
+      perlinNoise[j] = perlinNoise[j] || 0;
+      perlinNoise[j] += smoothNoiseList[i][j] * amplitude;
+    }
+  }
+  // normalization
+  for (i = 0; i < perlinNoise.length; ++i) {
+    perlinNoise[i] /= totalAmplitude;
+  }
+
+  return perlinNoise;
+
+  function generateSmoothNoise(octave: number) {
+    var noise = new Array(width * height);
+    var samplePeriod = Math.pow(2, octave);
+    var sampleFrequency = 1 / samplePeriod;
+    var noiseIndex = 0;
+    for (var y = 0; y < height; ++y) {
+      var sampleY0 = Math.floor(y / samplePeriod) * samplePeriod;
+      var sampleY1 = (sampleY0 + samplePeriod) % height;
+      var vertBlend = (y - sampleY0) * sampleFrequency;
+      for (var x = 0; x < width; ++x) {
+        var sampleX0 = Math.floor(x / samplePeriod) * samplePeriod;
+        var sampleX1 = (sampleX0 + samplePeriod) % width;
+        var horizBlend = (x - sampleX0) * sampleFrequency;
+
+        // blend top two corners
+        var top = interpolate(whiteNoise[sampleY0 * width + sampleX0], whiteNoise[sampleY1 * width + sampleX0], vertBlend);
+        // blend bottom two corners
+        var bottom = interpolate(whiteNoise[sampleY0 * width + sampleX1], whiteNoise[sampleY1 * width + sampleX1], vertBlend);
+        // final blend
+        noise[noiseIndex] = interpolate(top, bottom, horizBlend);
+        noiseIndex += 1;
+      }
+    }
+    return noise;
+  }
+}
+function generateWhiteNoise(width: number, height: number) {
+  var noise = new Array(width * height);
+  for (var i = 0; i < noise.length; ++i) {
+    noise[i] = Math.random();
+  }
+  return noise;
+}
+function interpolate(x0: number, x1: number, alpha: number) {
+  return x0 * (1 - alpha) + alpha * x1;
+}
+/* PERLIN NOISE */
